# 2_API/openapi-specification.yaml
openapi: 3.0.3
info:
  title: iWORKZ Platform API
  description: |
    Comprehensive API specification for the iWORKZ AI-powered employment platform.
    This API provides endpoints for talent matching, compliance management, and 
    blockchain-verified credential management specifically designed for the Japanese market.
  version: 1.0.0
  contact:
    name: iWORKZ API Support
    email: api-support@iworkz.jp
    url: https://docs.iworkz.jp
  license:
    name: Proprietary
    url: https://iworkz.jp/license
servers:
  - url: https://api.iworkz.jp/v1
    description: Production server
  - url: https://staging-api.iworkz.jp/v1
    description: Staging server
  - url: http://localhost:3000/v1
    description: Development server
security:
  - BearerAuth: []
  - ApiKeyAuth: []
paths:
...
  /jobs:
    get:
      tags: [Jobs]
      summary: Search and list jobs
      operationId: searchJobs
      parameters:
        - name: q
          in: query
          description: Search query
          schema:
            type: string
        - name: location
          in: query
          description: Job location
          schema:
            type: string
        - name: skills
          in: query
          description: Required skills (comma-separated)
          schema:
            type: string
        - name: salary_min
          in: query
          description: Minimum salary
          schema:
            type: integer
        - name: page
          in: query
          description: Page number
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          description: Items per page
          schema:
            type: integer
            default: 20
            maximum: 100
      responses:
        '200':
          description: Jobs retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  jobs:
                    type: array
                    items:
                      $ref: '#/components/schemas/Job'
                  pagination:
                    $ref: '#/components/schemas/Pagination'
                  total:
                    type: integer

    post:
      tags: [Jobs]
      summary: Create new job posting
      operationId: createJob
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/JobCreate'
      responses:
        '201':
          description: Job created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Job'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
  /credentials/{credentialId}/verify:
    get:
      tags: [Credentials]
      summary: Verify credential authenticity
      operationId: verifyCredential
      parameters:
        - name: credentialId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Credential verification result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CredentialVerification'

  # Analytics Endpoints
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
  schemas:
...
    Job:
      type: object
      properties:
        id:
          type: string
          format: uuid
        title:
          type: string
        company:
          $ref: '#/components/schemas/Company'
        location:
          $ref: '#/components/schemas/Location'
        description:
          type: string
        requirements:
          type: array
          items:
            type: string
        skills:
          type: array
          items:
            $ref: '#/components/schemas/Skill'
        salary:
          $ref: '#/components/schemas/Salary'
        employment_type:
          type: string
          enum: [full_time, part_time, contract, temporary]
        remote_work:
          type: boolean
        posted_at:
          type: string
          format: date-time
        expires_at:
          type: string
          format: date-time
        status:
          type: string
          enum: [active, paused, closed, expired]
    JobCreate:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        requirements:
          type: array
          items:
            type: string
        skills:
          type: array
          items:
            type: string
        location:
          $ref: '#/components/schemas/LocationInput'
        salary:
          $ref: '#/components/schemas/SalaryInput'
        employment_type:
          type: string
          enum: [full_time, part_time, contract, temporary]
        remote_work:
          type: boolean
          default: false
        visa_sponsorship:
          type: boolean
          default: false
      required: [title, description, employment_type]

    # Matching Schemas
...
    Location:
      type: object
      properties:
        address:
          type: string
        city:
          type: string
        prefecture:
          type: string
        postal_code:
          type: string
        country:
          type: string
          default: JP
        coordinates:
          type: object
          properties:
            lat:
              type: number
            lng:
              type: number
    Skill:
      type: object
      properties:
        name:
          type: string
        level:
          type: string
          enum: [beginner, intermediate, advanced, expert]
        years_experience:
          type: integer
        certified:
          type: boolean
    Company:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        industry:
          type: string
        size:
          type: string
          enum: [startup, small, medium, large, enterprise]
        website:
          type: string
          format: uri
    Salary:
      type: object
      properties:
        min:
          type: integer
        max:
          type: integer
        currency:
          type: string
          default: JPY
        period:
          type: string
          enum: [hourly, daily, monthly, yearly]
    Pagination:
      type: object
      properties:
        page:
          type: integer
        limit:
          type: integer
        total_pages:
          type: integer
        has_next:
          type: boolean
        has_previous:
          type: boolean
    Error:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
        code:
          type: string
        details:
          type: object
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...